Remove DNS sections, start instance last
My Windows instance is now self-updating its DNS after it loads via a PowerShell script. This is a little slower (DNS isn't updated until after boot) but centralizes the consern of DNS updating so that I can start the instance from places other than this script and ensure that the DNS is updated -- could actually do the same with the volume snapshotting.

This also now snapshots the volume before starting the instance -- unsure why I didn't always do it that way.